Implement base class for k-clustering algorithms
