+++ heat/cluster/__init__.py
@@ -1,2 +1,3 @@
+from .kcluster import *
 from .kmeans import *
 from .spectral import *
